Fase 2: Implementación de excepción NumeroDebeSerPositivo e ingrese_numero con manejo de excepciones
